Improvement/documentation correction (#21)
* Minor documentation update

When using `composer exec <command>` with options, you have to include `--` to let composer know to pass the options through. This isn't well documented within composer itself, but is mentioned in the terminal help listing for `composer exec` and mentioned on other, similar commands for composer.

See:
- `composer exec --help`
- https://getcomposer.org/doc/articles/scripts.md#running-scripts-manually (not exec, but it follows the same rules)

* Markdown compatibility change

Making this table conform to more standardized Markdown. Doesn't affect GitHub, but could cause the table to render incorrectly in some parsers.